cdk/requirements-cdk.txt
aws-cdk.core==1.72.0
aws-cdk.aws-ec2==1.72.0
aws-cdk.aws-elasticache==1.72.0
aws-cdk.aws-iam==1.72.0
aws-cdk.aws-logs==1.72.0
aws-cdk.aws-s3==1.72.0
aws-cdk.aws-ecr==1.72.0
aws-cdk.aws-ecs==1.72.0
aws-cdk.aws-ecs-patterns==1.72.0
aws-cdk.aws-applicationautoscaling==1.72.0

boto3==1.15.16
botocore==1.18.16
